move old components and create files for new ones
