+++ frontend/app/components/Calendar.tsx
@@ -1,0 +1,3 @@
+export function Calendar() {
+  return <div></div>;
+}
