+++ frontend/app/components/DegreeAudit.tsx
@@ -1,0 +1,3 @@
+export function DegreeAudit() {
+  return <div></div>;
+}
