+++ frontend/app/components/old/audit/DegreeAuditPart.tsx
