+++ frontend/app/components/old/schedule/SlideBar.tsx
