+++ frontend/app/routes.ts
@@ -1,12 +1,7 @@
 import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
-  index("routes/homePage.tsx"),
-  route("/dashboard", "routes/dashboard.tsx"),
-  
+  index("routes/home.tsx"),
   route("/login", "routes/loginPage.tsx"),
-  route("/plans", "routes/planPages.tsx"),
-  route("/map", "routes/mapPage.tsx"),
-  route("/chatbot", "routes/chatbotPage.tsx"),
-  route("/classes", "routes/classesPage.tsx"),
+  route("/dashboard", "routes/dashboard.tsx"),
 ] satisfies RouteConfig;
